cryptora/requirements/local.txt
# Local development dependencies go here
-r test.txt
django-debug-toolbar==1.8
Sphinx==1.6.2
